+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Expected: May 2026
 
 * Fixed: [Custom Expansion Functions called twice on Pressing TAB](https://github.com/clicon/cligen/issues/92)
 * Fixed: [Validation of Partial match](https://github.com/clicon/cligen/issues/134)
+* Fixed: [Inconsistent error messages for out of range values](https://github.com/clicon/clixon/issues/567)
 
 ## 7.7.0
 21 February 2026